+++ src/components/ExperienceDetail/ChartsZone/index.js
@@ -1,12 +1,20 @@
-import React from 'react';
+import React, { useEffect, useMemo } from 'react';
 import PropTypes from 'prop-types';
 import cn from 'classnames';
 import R from 'ramda';
+import { useDispatch, useSelector } from 'react-redux';
 import { PageType, generatePageURL } from 'constants/companyJobTitle';
 import ChartWrapper from '../../LandingPage/ChartWrapper';
 import loadable from '@loadable/component';
 import styles from '../../LandingPage/SummarySection.module.css';
 import moduleStyles from './ChartsZone.module.css';
+import { queryCompanyOverviewStatistics } from 'actions/company';
+import { queryJobTitleOverviewStatistics } from 'actions/jobTitle';
+import {
+  companyOverviewStatisticsBoxSelectorByName,
+  jobTitleOverviewStatisticsBoxSelectorByName,
+} from 'selectors/companyAndJobTitle';
+import { isFetched } from 'utils/fetchBox';
 
 const isEmptyOrNull = R.either(R.isEmpty, R.isNil);
 
@@ -17,14 +25,41 @@ const JobTitleDistributionChart = loadable(() =>
   import('common/Charts/JobTitleDistributionChart'),
 );
 
-const ChartsZone = ({
-  experience: {
-    company: { name: companyName },
-    job_title: { name: jobTitle },
-  },
-  jobAverageSalaries,
-  salaryDistributionBins,
-}) => {
+const ChartsZone = ({ companyName, jobTitle }) => {
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(queryCompanyOverviewStatistics(companyName));
+  }, [dispatch, companyName]);
+
+  useEffect(() => {
+    dispatch(queryJobTitleOverviewStatistics(jobTitle));
+  }, [dispatch, jobTitle]);
+
+  const jobAverageSalaries = useSelector(
+    useMemo(
+      () => state => {
+        const box = companyOverviewStatisticsBoxSelectorByName(companyName)(
+          state,
+        );
+        return isFetched(box) && box.data ? box.data.jobAverageSalaries : [];
+      },
+      [companyName],
+    ),
+  );
+
+  const salaryDistributionBins = useSelector(
+    useMemo(
+      () => state => {
+        const box = jobTitleOverviewStatisticsBoxSelectorByName(jobTitle)(
+          state,
+        );
+        return isFetched(box) && box.data ? box.data.salaryDistribution : [];
+      },
+      [jobTitle],
+    ),
+  );
+
   if (
     isEmptyOrNull(jobAverageSalaries) &&
     isEmptyOrNull(salaryDistributionBins)
@@ -70,9 +105,8 @@ const ChartsZone = ({
 };
 
 ChartsZone.propTypes = {
-  experience: PropTypes.object.isRequired,
-  jobAverageSalaries: PropTypes.array.isRequired,
-  salaryDistributionBins: PropTypes.array.isRequired,
+  companyName: PropTypes.string.isRequired,
+  jobTitle: PropTypes.string.isRequired,
 };
 
 export default ChartsZone;
